history: code refactoring for sharing code needed to undo/redo styles.
